all: restructure tos; include PID in the default process TLS
The discrepancy in the size of `long' between Plan 9 and
UNIX causes miscalculation of the offset of `pid' from Tos,
in programs that must explicitly calculate that, such as
the Go runtime. Thus, we include it in the default TLS for
all processes so that the Go runtime can access it that
way, and move it to the "top of stack" so as to resolve
the offset miscalculation problem in the future (and also
to preserve the place of the Go pointers in TLS, for the Go
runtime). Moreover, we explicitly make it `uvlong' so as
to support 64-bit PIDs.

Reviewed-on: http://nxm.coreboot.org/review/156
Tested-by: Jenkins QA

+++ sys/src/nxm/k10/trap.c
@@ -303,7 +303,7 @@ kexit(Ureg*)
 	/* Set thread local storage, used now in both Linux
 	 * and Go processes. 
 	 */
-	wrmsr(FSbase, (ulong) up->tls);
+	wrmsr(FSbase, (uintptr) up->tls);
 }
 
 void
